# bank_rest-main/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bank Cards API
  version: v1
  description: Операции по картам и аутентификация JWT
servers:
  - url: http://localhost:8081
paths:
  /api/token:
    post:
      tags: [Auth]
      summary: Выдать JWT-токен
      description: Аутентифицирует пользователя по username/password и возвращает JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenResponse' }
        '401':
          description: Неверные креденшилы
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /api/cards/admin:
    get:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
      summary: Список карт (ADMIN)
      parameters:
        - in: query
          name: page
          schema: { type: integer, minimum: 0, default: 0 }
        - in: query
          name: size
          schema: { type: integer, minimum: 1, default: 20 }
        - in: query
          name: sort
          schema: { type: string, example: "id,desc" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PageCardResponse' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  /api/cards/admin/{id}:
    delete:
      tags: [Cards]
      security: [{ bearerAuth: [] }]
      summary: Удалить карту по id (ADMIN)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '204': { description: No Content }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '403':
          description: Forbidden
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: Not Found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string, example: admin }
        password: { type: string, example: 123456 }

    TokenResponse:
      type: object
      required: [token]
      properties:
        token: { type: string, example: eyJhbGciOiJIUzI1NiJ9... }

    ErrorResponse:
      type: object
      properties:
        message: { type: string, example: Internal server error }
        detailedMessage: { type: string, example: Stacktrace or details... }
        localDateTime: { type: string, example: 2025-11-10T13:42:23.193502923 }

    CardResponseDto:
      type: object
      properties:
        id: { type: integer, format: int64, example: 1 }
        maskedNumber: { type: string, example: "4111 **** **** 1111" }
        holderName: { type: string, example: "IVAN IVANOV" }
        brand: { type: string, example: "VISA" }
        status: { type: string, example: "ACTIVE" }
        validUntil: { type: string, example: "12/2027" }

    PageCardResponse:
      type: object
      properties:
        content:
          type: array
          items: { $ref: '#/components/schemas/CardResponseDto' }
        number: { type: integer, example: 0 }
        size: { type: integer, example: 20 }
        totalElements: { type: integer, example: 123 }
        totalPages: { type: integer, example: 7 }
        first: { type: boolean, example: true }
        last: { type: boolean, example: false }